+++ modules-services/prov-service-security-pac/src/main/java/org/openprovenance/prov/service/security/pac/ClientConfiguration.java
@@ -1,4 +1,4 @@
-package org.openprovenance.prov.service.security;
+package org.openprovenance.prov.service.security.pac;
 
 import org.pac4j.core.client.Client;