+++ modules-services/prov-service-security-pac/src/main/java/org/openprovenance/prov/service/security/pac/RoleAuthorizationGenerator.java
@@ -1,4 +1,4 @@
-package org.openprovenance.prov.service.security;
+package org.openprovenance.prov.service.security.pac;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;