+++ RELEASES.md
@@ -49,6 +49,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * docker-service-templates: a secure docker image for templates service
 * prov-template-archetype: skeleton
 * prov-jsonld: fixed incompatibility issues with jsonld1.1
+* prov-service-security-pac: created sub-module
 
 `2.0.5`
 * prov-template-library: cleaned up pom.xml